+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 5.5.1 - 2026-09-25
 
 - Add Custom Model field type (requires MB Custom Table)
+- Add Model object type to relationships UI
 
 ### 5.5.0 - 2026-09-17